[MIG] pos_access_right: Migration to 16.0 - revision
[MIG] pos_access_right: Migration to 16.0 - revision 2

[MIG] pos_access_right: Migration to 16.0 - adjusted test unit and add one line information in description.

[MIG] pos_access_right: Migration to 16.0 - repair test unit error.

[MIG] pos_access_right: Migration to 16.0 - still repair test unit error.

[MIG] pos_access_right: Migration to 16.0 - give access right for backspace button in keyboard and button remove in order line to delete order line.

+++ pos_access_right/static/src/js/TicketScreen.js
@@ -7,11 +7,16 @@ odoo.define("pos_access_right.TicketScreen", function (require) {
     const PosTicketScreen = (TicketScreen) =>
         class extends TicketScreen {
             get hasNewOrdersControlRights() {
-                return this.env.pos.get_cashier().hasGroupMultiOrder;
+                if (this.env.pos.config.module_pos_hr)
+                    return this.env.pos.get_cashier().hasGroupMultiOrder;
+                return this.env.pos.user.hasGroupMultiOrder;
             }
 
             async _onDeleteOrder({detail: order}) {
-                if (this.env.pos.get_cashier().hasGroupDeleteOrder) {
+                if (
+                    this.env.pos.get_cashier().hasGroupDeleteOrder &&
+                    this.env.pos.user.hasGroupDeleteOrder
+                ) {
                     return super._onDeleteOrder({detail: order});
                 }
                 return false;
